+++ Assets/com.gamedeveducation.aibehaviours/Runtime/HybridGOAP/Scripts/Goals/Common/GOAPGoal_Idle.cs
@@ -1,0 +1,14 @@
+using HybridGOAP;
+using UnityEngine;
+
+namespace HybridGOAP
+{
+    [AddComponentMenu("AI/GOAP/Goals/Goal: Idle Goal")]
+    public class GOAPGoal_Idle : GOAPGoalBase
+    {
+        public override void PrepareForPlanning()
+        {
+            Priority = GoalPriority.Minimum + 1;
+        }
+    }
+}