+++ Assets/com.gamedeveducation.aibehaviours/Runtime/HybridGOAP/Scripts/Goals/Common/GOAPGoal_Wander.cs
@@ -1,0 +1,14 @@
+using HybridGOAP;
+using UnityEngine;
+
+namespace HybridGOAP
+{
+    [AddComponentMenu("AI/GOAP/Goals/Goal: Wander Goal")]
+    public class GOAPGoal_Wander : GOAPGoalBase
+    {
+        public override void PrepareForPlanning()
+        {
+            Priority = GoalPriority.Ambient;
+        }
+    }
+}