+++ Assets/com.gamedeveducation.aibehaviours/Runtime/HybridGOAP/TestScenes/GOAP_StateMachines/SimpleGOAPBrain_SM.cs
@@ -1,0 +1,13 @@
+namespace HybridGOAP
+{
+    public class SimpleGOAPBrain_SM : GOAPBrainBase
+    {
+        protected override void ConfigureBlackboard()
+        {
+        }
+
+        protected override void ConfigureBrain()
+        {
+        }
+    }
+}